refactor(within-py): split lib.rs into role modules
Split the ~900-line lib.rs into config / results / convert / api / objects
modules; lib.rs now holds only the #[pymodule] registration and module
declarations. The _within module name and every Python-visible class,
function, and signature are unchanged.

Also from the audit:
- into_py_batch_result returns PyResult and maps the result-shape-invariant
  check through value_err, so a violation surfaces as a clean ValueError
  instead of an opaque PanicException across the FFI boundary.
- solve_batch performs its O(n_obs * n_rhs) response-column copy inside
  py.allow_threads, releasing the GIL during the copy.
